✨ Lifting State Up, Again
