+++ tic-tac-toe/src/Board.js
@@ -1,37 +1,12 @@
-import { useState } from 'react'
-import calculateWinner from './calculateWinner'
 import Square from './Square'
 
 function Board(props) {
-  const [state, setState] = useState({
-    squares: Array(9).fill(null),
-    xIsNext: true,
-  })
-
-  const handleClick = (i) => {
-    const squares = state.squares.slice()
-    if (calculateWinner(squares) || squares[i]) {
-      return
-    }
-    squares[i] = state.xIsNext ? 'X' : 'O'
-    setState({ squares: squares, xIsNext: !state.xIsNext })
-  }
-
   const renderSquare = (i) => {
-    return <Square value={state.squares[i]} onClick={() => handleClick(i)} />
-  }
-
-  const winner = calculateWinner(state.squares)
-  let status
-  if (winner) {
-    status = 'Winner: ' + winner
-  } else {
-    status = 'Next player: ' + (state.xIsNext ? 'X' : 'O')
+    return <Square value={props.squares[i]} onClick={() => props.onClick(i)} />
   }
 
   return (
     <div>
-      <div className="status">{status}</div>
       <div className="board-row">
         {renderSquare(0)}
         {renderSquare(1)}
